Update AWS Lambda Runtime
The AWS Lambda that can be used to start the server after an auto-
shutdown uses the default runtime from AWS Lambda. With our current CDK
version this is NodeJS 14.0, which is not supported by AWS anymore.

To resolve, first update the CDK lib version (in fact we updated all
dependencies since some were years out of date from our original fork),
and force NodeJS 22.0 to be used.