Update Ec2 Instance Type
Currently the instance type is set to `m5a.large`, which is smaller than
the recommended Satisfactory ec2 server specs from [here][1].

Instead, update to `m6a.xlarge`, which meets the minimum dedicated
server specifications.

Furthermore I've updated the script to allow specifying the instance
type from the Config. It is currently hard-coded into
`server-hosting-stack.ts`. Putting it into the Config should allow for
easy updating in future if the server size needs to be increased.

[1]: https://satisfactory.wiki.gg/wiki/Dedicated_servers#Requirements