Emit Public IP When Connecting to API Endpoint
The people connecting to the server start API Gateway endpoint will not
know the public IP address to connect to.

Instead of returning basic JSON, return a full HTML page which includes
the public IP of the server, as well as basic instructions on how to
connect.

An HTML template must now be provided by the user to provide to the user
in the API Gateway response.

I've also updated the README quite a bit to reflect the current reality
of the repo.